% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-nodes.R
\name{all_nodes}
\alias{all_nodes}
\title{Get all nodes.}
\usage{
all_nodes(edge_list, verbose = F)
}
\arguments{
\item{edge_list}{Edge list as a \code{data.table} which must contain the columns \code{Origin},
\code{Destination}, \code{OriginType}, and \code{DestinationType}.}

\item{verbose}{Should total node count be printed to the console?}
}
\value{
List of unique nodes.
}
\description{
Retrieve a list of all unique nodes from an edge list.
}
