% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-neighbors.R
\name{clean_mp}
\alias{clean_mp}
\title{Clean type name along meta-path.}
\usage{
clean_mp(type)
}
\arguments{
\item{type}{Name of the type along any meta-path of interest.}
}
\value{
Cleaned type name.
}
\description{
For a given meta-path and root node, when the meta-ath \code{data.table} is constructed, steps along
the meta-path with the same type (e.g., \code{Author-Date-Author}) are assigned the same column name.
Then, \code{make.unique()} is run to deduplicate column names by appending \code{.1}, \code{.2}, etc.
When querying the edge list (or neighbor list), the original (non-unique) type name is required.
Hence, this function strips the unique suffix from the type name (e.g., removes \code{.1}, \code{.2}, etc.).
}
