% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-sets.R
\name{compare_sets}
\alias{compare_sets}
\title{Compute meta-path based similarity scores between two sets of nodes.}
\usage{
compare_sets(
  set1,
  set2,
  mp,
  metric = c("pc", "pathsim", "npc", "dwpc"),
  method = c("shortest"),
  node_list,
  edge_list = NULL,
  neighbor_list = NULL,
  verbose = TRUE
)
}
\arguments{
\item{set1}{IDs of node set #1 as a vector.}

\item{set2}{IDs of node set #2 as a vector.}

\item{mp}{Meta-path as a vector of node types (e.g., \code{c("TypeA", "TypeB", "TypeC")}).}

\item{metric}{A vector of permissible similarity metrics (e.g., \code{c("npc", "dwpc")}), see defined metrics below:
\describe{
  \item{Path Count}{Use \code{"pc"}.}
  \item{Normalized Path Count}{Use \code{"npc"}.}
  \item{Degree-Weighted Path Count}{Use \code{"dwpc"}.}
  \item{PathSim}{Use \code{"pathsim"}.}}}

\item{node_list}{Node list as a \code{data.table} which must contain the columns \code{Node}
(corresponding to either \code{Origin} or \code{Destination} in the edge list), and \code{Node Type}
(corresponding to either \code{OriginType} or \code{DestinationType} in the edge list). To disable type checking,
set \code{node_list} to \code{"NoChecking"}.}

\item{neighbor_list}{Neighbor reference object constructed by \code{get_neighbor_list()}.}

\item{verbose}{Should the intermediate calculations be printed to the console?}

\item{reference_list}{Either an edge list as a \code{data.table} which must contain the columns \code{Origin},
\code{Destination}, \code{OriginType}, and \code{DestinationType}, or a neighbor reference object constructed
by \code{get_neighbor_list()}.}

\item{list_type}{If an edge list is provided, specify \code{"edge"}. If a neighbor list is provided, specify \code{"neighbor"}.}
}
\value{
A list with six elements:
\describe{
  \item{Set1}{IDs of node set #1 provided (i.e., \code{set1}).}
  \item{Set2}{ID of node set #2 provided (i.e., \code{set2}).}
  \item{MP}{Meta-path provided (i.e., \code{mp}).}
  \item{Comparisons}{List of pairwise comparisons.}
  \item{Details}{Intermediate computations produced by \code{get_similarity()}.}
  \item{Similarities}{Pairwise similarity scores.}
  \item{SetSimilarity}{Aggregate set similarity score(s).}}
}
\description{
Given a specified meta-path, similarity metric(s), and aggregation method(s), compute the meta-path based similarity score between two nodes.
}
\seealso{
\code{get_neighbor_list()} for neighbor reference object construction.

\code{get_neighbor_list()} for neighbor reference object construction.

\code{get_neighbor_list()} for neighbor reference object construction,
\code{get_similarity_function()} for similarity metrics, and
\code{get_aggregation_function()} for aggregation methods. This function is a
wrapper around \code{get_similarity()}.
}
