% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-sets.R
\name{extract_sim}
\alias{extract_sim}
\title{Extract similarity scores from pairwise comparisons.}
\usage{
extract_sim(x)
}
\arguments{
\item{x}{A \code{data.table} containing the output of a \code{get_similarity()$Similarity} call.}

\item{verbose}{Should aggregation method be printed to the console?}
}
\value{
A data.table with multiple columns elements:
\describe{
  \item{Origin}{The ID of the origin node.}
  \item{Destination}{The ID of the destination node.}
  \item{Metric}{The name of the similarity metric used (e.g., \code{"PathSim"}).}
  \item{Method}{The name of the aggregation method used.}
  \item{Similarity}{The pairwise meta-path based similarity score.}}
}
\description{
During aggregation of meta-path based similarity, extract similarity scores from pairwise aggregations.
}
\seealso{
\code{get_similarity()} for pairwise meta-path based similarity calculation.
}
