% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-methods.R
\name{get_aggregation_function}
\alias{get_aggregation_function}
\title{Retrieve aggregation method.}
\usage{
get_aggregation_function(method_name, get_verbose, ...)
}
\arguments{
\item{method_name}{Readable description of the desired aggregation method:
\describe{
  \item{Shortest}{Use \code{"shortest"}.}}}

\item{get_verbose}{Should output be printed to console?}

\item{...}{Other arguments to pass to appropriate meta-path similarity function.}
}
\value{
The output of the desired aggregation method.
}
\description{
For a desired aggregation method, retrieve the function which aggregates the meta-path based similarity scores.
}
\seealso{
See the following aggregation functions:
\describe{
  \item{Shortest}{\code{aggregate_shortest()}}}
}
