% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-metrics.R
\name{get_dwpc}
\alias{get_dwpc}
\title{Compute degree-weighted path count similarity.}
\usage{
get_dwpc(
  x,
  y,
  paths_x,
  paths_y,
  reference_list,
  list_type = c("edge", "neighbor"),
  verbose = TRUE,
  w = 0.4
)
}
\arguments{
\item{x}{ID of the origin node.}

\item{y}{ID of the destination node.}

\item{paths_x}{Paths from the origin node following the meta-path of interest as a \code{data.table}.}

\item{paths_y}{Paths from the destination node following the meta-path of interest as a \code{data.table}.}

\item{reference_list}{Either an edge list as a \code{data.table} which must contain the columns \code{Origin},
\code{Destination}, \code{OriginType}, and \code{DestinationType}, or a neighbor reference object constructed
by \code{get_neighbor_list()}.}

\item{list_type}{If an edge list is provided, specify \code{"edge"}. If a neighbor list is provided, specify \code{"neighbor"}.}

\item{verbose}{Should the intermediate calculations be printed to the console?}
}
\value{
A list with two elements:
\describe{
  \item{Metric}{The name of the similarity metric (i.e., \code{"Degree-Weighted Path Count"}).}
  \item{Similarity}{The degree-weighted path count similarity score.}}
}
\description{
Use the degree-weighted path count metric to compute a meta-path based similarity score. Node that, in this implementation,
type-specific degrees are used (except for the last step of the meta-path).
}
\references{
\cite{Himmelstein, D. S. & Baranzini, S. E. Heterogeneous Network Edge Prediction: A Data Integration Approach to Prioritize Disease-Associated Genes. PLOS Computational Biology 11, e1004259 (2015).}
}
\seealso{
\code{get_neighbor_list()} for neighbor reference object construction.

Other similarity: 
\code{\link{get_npc}()},
\code{\link{get_pathsim}()},
\code{\link{get_pc}()}
}
\concept{similarity}
