% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-neighbor-list.R
\name{get_neighbor_list}
\alias{get_neighbor_list}
\title{Construct neighbors reference object.}
\usage{
get_neighbor_list(
  edge_list,
  file_path = NULL,
  intermediate_file_path = NULL,
  verbose = F
)
}
\arguments{
\item{edge_list}{Edge list as a \code{data.table} which must contain the columns \code{Origin},
\code{Destination}, \code{OriginType}, and \code{DestinationType}.}

\item{file_path}{File path to save final \code{data.table} reference object. If no path is specified, the
reference object is not saved.}

\item{intermediate_file_path}{File path to save intermediate list of lists reference object. If no path is
specified, the intermediate reference object is not saved.}

\item{verbose}{Should node types and total node count be printed to the console?}
}
\value{
Neighbors reference object as a \code{data.table}
}
\description{
For a given edge list, construct a nested list (i.e., dictionary of dictionaries in Python syntax)
of the neighbors of each node in the edge list stratified by type. This reference object can be
used instead of the edge list in subsequent \code{metapaths} functions, since searching the edge list
to recompute neighbors each time would have \code{O(n)} complexity, while searching a pre-computed neighbors
dictionary requires a fixed amount of time.
}
\examples{
mtcars_neighbor_list = get_neighbor_list(mtcars_edge_list, verbose = T)
head(mtcars_neighbor_list)
}
