% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-neighbors.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get node neighbors of specified type.}
\usage{
get_neighbors(root, type, eslist)
}
\arguments{
\item{root}{ID of the root node.}

\item{type}{Desired type of the neighbors.}

\item{eslist}{Edge list as a \code{data.table} which must contain the columns \code{subject},
\code{object}, \code{subject_type}, and \code{object_type}.}
}
\value{
List of neighbors.
}
\description{
Retrieve the neighbors of a given node which are of a given type.
}
\author{
Ayush Noori
}
