% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-neighbors.R
\name{get_neighbors_type}
\alias{get_neighbors_type}
\title{Get all node neighbors by type.}
\usage{
get_neighbors_type(root, edge_list_types, edge_list)
}
\arguments{
\item{root}{ID of the root node.}

\item{edge_list_types}{Specified node types to stratify by.}

\item{edge_list}{Edge list as a \code{data.table} which must contain the following columns:
\describe{
  \item{\code{Origin}}{IDs of the origin nodes for each edge.}
  \item{\code{Destination}}{IDs of the destination nodes for each edge.}
  \item{\code{OriginType}}{Types of the origin node for each edge.}
  \item{\code{DestinationType}}{Types of the destination node for each edge.}
  \item{\code{EdgeType}}{Types of each edge.}}}
}
\value{
Nested list of neighbors grouped by type.
}
\description{
Retrieve all neighbors of a given node, stratified by specific node types.
}
