% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-metrics.R
\name{get_pathsim}
\alias{get_pathsim}
\title{Compute PathSim similarity.}
\usage{
get_pathsim(
  x,
  y,
  paths_x,
  paths_y,
  reference_list = NULL,
  list_type = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{ID of the origin node.}

\item{y}{ID of the destination node.}

\item{paths_x}{Paths from the origin node following the meta-path of interest as a \code{data.table}.}

\item{paths_y}{Paths from the destination node following the meta-path of interest as a \code{data.table}.}

\item{reference_list}{Either an edge list as a \code{data.table} which must contain the columns \code{Origin},
\code{Destination}, \code{OriginType}, and \code{DestinationType}, or a neighbor reference object constructed
by \code{get_neighbor_list()}.}

\item{list_type}{If an edge list is provided, specify \code{"edge"}. If a neighbor list is provided, specify \code{"neighbor"}.}

\item{verbose}{Should the intermediate calculations be printed to the console?}
}
\value{
A list with two elements:
\describe{
  \item{Metric}{The name of the similarity metric (i.e., \code{"PathSim"}).}
  \item{Similarity}{The PathSim similarity score.}}
}
\description{
Use the PathSim metric to compute a meta-path based similarity score. Please note that PathSim can ONLY
be used for symmetric meta-paths.
}
\references{
\cite{Sun, Y., Han, J., Yan, X., Yu, P. S. & Wu, T. PathSim: meta path-based top-K similarity search in heterogeneous information networks. Proc. VLDB Endow. 4, 992–1003 (2011).}
}
\seealso{
\code{get_neighbor_list()} for neighbor reference object construction.

Other similarity: 
\code{\link{get_dwpc}()},
\code{\link{get_npc}()},
\code{\link{get_pc}()}
}
\concept{similarity}
