% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-metrics.R
\name{get_similarity_function}
\alias{get_similarity_function}
\title{Retrieve similarity score metric.}
\usage{
get_similarity_function(metric_name, get_verbose, ...)
}
\arguments{
\item{metric_name}{Readable description of the desired similarity score:
\describe{
  \item{Path Count}{Use \code{"pc"}.}
  \item{Normalized Path Count}{Use \code{"npc"}.}
  \item{Degree-Weighted Path Count}{Use \code{"dwpc"}.}
  \item{PathSim}{Use \code{"pathsim"}.}}}

\item{get_verbose}{Should output be printed to console?}

\item{...}{Other arguments to pass to appropriate meta-path similarity function.}
}
\value{
A list with two elements:
\describe{
  \item{Metric}{The human readable name of the similarity metric.}
  \item{Similarity}{The appropriate similarity score.}}
}
\description{
For a desired meta-path based similarity metric, retrieve the function which computes the appropriate similarity score, then compute similarity.
}
\seealso{
See the following functions to compute various similarity scores:
\describe{
  \item{Path Count}{\code{get_pc()}}
  \item{Normalized Path Count}{\code{get_npc()}}
  \item{Degree-Weighted Path Count}{\code{get_dwpc()}}
  \item{PathSim}{\code{get_pathsim()}}}
}
