% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-sets.R
\name{map_log}
\alias{map_log}
\title{Add dividers between analysis logs.}
\usage{
map_log(res, verbose)
}
\arguments{
\item{res}{An intermediate value of an iterable function (e.g., \code{purrr::map}).}

\item{verbose}{Should dividers be printed to the console?}
}
\value{
\code{res} is returned as is.
}
\description{
Print dividers to the console between analysis logs when \code{verbose} is \code{TRUE}.
}
