% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-neighbors.R
\name{map_neighbors}
\alias{map_neighbors}
\title{Step along meta-path to retrieve neighbors.}
\usage{
map_neighbors(
  roots_subset,
  current_mp,
  next_mp,
  reference_list,
  list_type = c("edge", "neighbor")
)
}
\arguments{
\item{roots_subset}{Paths from a specific root node as a \code{data.table}.}

\item{current_mp}{Label of the current type along the meta-path.}

\item{next_mp}{Label of the next type along the meta-path.}

\item{reference_list}{Either an edge list as a \code{data.table} which must contain the columns \code{Origin},
\code{Destination}, \code{OriginType}, and \code{DestinationType}, or a neighbor reference object constructed
by \code{get_neighbor_list()}.}

\item{list_type}{If an edge list is provided, specify \code{"edge"}. If a neighbor list is provided, specify \code{"neighbor"}.}
}
\value{
Neighbors of the root nodes of a specified type (along the meta-path) as a \code{data.table}.
}
\description{
For a given set of root nodes, step along a given meta-path to find their neighbors
using either \code{get_neighbors()} (for edge lists) or \code{search_neighbors()} (for
neighbor lists).
}
\seealso{
\code{get_neighbor_list()} for neighbor reference object construction.
}
