% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapaths-package.R
\docType{package}
\name{metapaths-package}
\alias{metapaths-package}
\alias{metapaths}
\title{Meta-path based network similarity.}
\description{
This package facilitates the computation of meta-path based network similarity metrics in R.
}
\references{
\url{https://www.ayushnoori.com/metapaths}
}
\author{
Ayush Noori \email{anoori@college.harvard.edu}

Amelia Tan \email{amelia_tan@hms.harvard.edu}

Michelle M. Li \email{michelleli@g.harvard.edu}

Marinka Zitnik \email{marinka@hms.harvard.edu}
}
