% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-types.R
\name{node_types}
\alias{node_types}
\title{Get node types.}
\usage{
node_types(edge_list, verbose = F)
}
\arguments{
\item{edge_list}{Edge list as a \code{data.table} which must contain the columns \code{Origin},
\code{Destination}, \code{OriginType}, and \code{DestinationType}.}

\item{verbose}{Should node types be printed to the console?}
}
\value{
List of node types.
}
\description{
Retrieve the node types from an edge list.
}
