% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-degrees.R
\name{search_degrees}
\alias{search_degrees}
\title{Get degree count by type.}
\usage{
search_degrees(
  root,
  type = NA,
  reference_list,
  list_type = c("edge", "neighbor"),
  consider_edge_type = TRUE
)
}
\arguments{
\item{root}{ID of the root node.}

\item{type}{Desired type of the neighbors.}

\item{reference_list}{Either an edge list as a \code{data.table} or a neighbor reference object constructed
by \code{get_neighbor_list()}. If an edge list is provided, the \code{data.table} must contain the following columns:
\describe{
  \item{\code{Origin}}{IDs of the origin nodes for each edge.}
  \item{\code{Destination}}{IDs of the destination nodes for each edge.}
  \item{\code{OriginType}}{Types of the origin node for each edge.}
  \item{\code{DestinationType}}{Types of the destination node for each edge.}
  \item{\code{EdgeType}}{Types of each edge.}}}

\item{list_type}{If an edge list is provided, specify \code{"edge"}. If a neighbor list is provided, specify \code{"neighbor"}.}

\item{consider_edge_type}{Should \code{EdgeType} be considered in the degree calculation? Only possible when using a edge list (i.e., \code{list_type} is \code{"edge"}) rather than a neighbor list.}
}
\value{
Degree count stratified by type.
}
\description{
Compute degree count stratified by node type using either an edge list or neighbor reference object computed by \code{get_neighbor_list()}.
}
\seealso{
\code{get_neighbor_list()} for neighbor reference object construction.
}
