% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-neighbors.R
\name{search_neighbors}
\alias{search_neighbors}
\title{Get node neighbors of specified type using neighbors reference object.}
\usage{
search_neighbors(root, type, neighbor_list)
}
\arguments{
\item{root}{ID of the root node.}

\item{type}{Desired type of the neighbors.}

\item{neighbor_list}{Neighbor reference object constructed by \code{get_neighbor_list()}.}
}
\value{
Vector of neighbors of a given type.
}
\description{
Retrieve the neighbors of a given node which are of a given type using the neighbors reference object.
}
\seealso{
[get_neighbor_list()] for neighbor reference object construction.
}
