% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-neighbors.R
\name{traverse_mp}
\alias{traverse_mp}
\title{Step along meta-path to get all conforming paths from root node.}
\usage{
traverse_mp(
  roots,
  mp,
  step = 1,
  reference_list,
  list_type = c("edge", "neighbor"),
  verbose = TRUE
)
}
\arguments{
\item{roots}{ID of the root node on the first iteration, otherwise, \code{data.table} of growing paths.}

\item{mp}{List of node types (i.e., a meta-path).}

\item{step}{Counter for recursive function.}

\item{reference_list}{Either an edge list as a \code{data.table} or a neighbor reference object constructed
by \code{get_neighbor_list()}. If an edge list is provided, the \code{data.table} must contain the following columns:
\describe{
  \item{\code{Origin}}{IDs of the origin nodes for each edge.}
  \item{\code{Destination}}{IDs of the destination nodes for each edge.}
  \item{\code{OriginType}}{Types of the origin node for each edge.}
  \item{\code{DestinationType}}{Types of the destination node for each edge.}
  \item{\code{EdgeType}}{Types of each edge.}}}

\item{list_type}{If an edge list is provided, specify \code{"edge"}. If a neighbor list is provided, specify \code{"neighbor"}.}

\item{verbose}{Should intermediate calculations be printed to console?}
}
\value{
List of all meta-paths from the root node following a specific meta-path, as a \code{data.table}.
}
\description{
Recursive function to traverse the meta-path until it is expounded. The variable \code{mp} is a list of node types (i.e., a meta-path),
which is shortened on each call of \code{traverse_mp()}.
}
\seealso{
\code{get_neighbor_list()} for neighbor reference object construction.
}
