% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-nodes.R
\name{all_nodes}
\alias{all_nodes}
\title{Get all nodes.}
\usage{
all_nodes(edge_list, verbose = F)
}
\arguments{
\item{edge_list}{Edge list as a \code{data.table} which must contain the following columns:
\describe{
  \item{\code{Origin}}{IDs of the origin nodes for each edge.}
  \item{\code{Destination}}{IDs of the destination nodes for each edge.}
  \item{\code{OriginType}}{Types of the origin node for each edge.}
  \item{\code{DestinationType}}{Types of the destination node for each edge.}
  \item{\code{EdgeType}}{Types of each edge.}}}

\item{verbose}{Should total node count be printed to the console?}
}
\value{
List of unique nodes.
}
\description{
Retrieve a list of all unique nodes from an edge list.
}
