% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-sets.R
\name{compare_sets}
\alias{compare_sets}
\title{Compute meta-path based similarity scores between two sets of nodes.}
\usage{
compare_sets(
  set1,
  set2,
  mp,
  metric = c("pc", "pathsim", "npc", "dwpc"),
  method = c("shortest"),
  node_list,
  edge_list = NULL,
  neighbor_list = NULL,
  check = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{set1}{IDs of node set #1 as a vector.}

\item{set2}{IDs of node set #2 as a vector.}

\item{mp}{Meta-path as a vector of node types (e.g., \code{c("TypeA", "TypeB", "TypeC")}).}

\item{metric}{A vector of permissible similarity metrics (e.g., \code{c("npc", "dwpc")}), see defined metrics below:
\describe{
  \item{Path Count}{Specify \code{"pc"}, corresponding to \code{get_pc()}.}
  \item{Normalized Path Count}{Specify \code{"npc"}, corresponding to \code{get_npc()}.}
  \item{Degree-Weighted Path Count}{Specify \code{"dwpc"}, corresponding to \code{get_dwpc()}.}
  \item{PathSim}{Specify \code{"pathsim"}, corresponding to \code{get_pathsim()}.}}}

\item{node_list}{Node list as a \code{data.table} which must contain the following columns:
\describe{
  \item{\code{Node}}{Node IDs (corresponding to either \code{Origin} or \code{Destination} in the edge list).}
  \item{\code{NodeType}}{Node types (corresponding to either \code{OriginType} or
  \code{DestinationType} in the edge list).}}}

\item{edge_list}{Edge list as a \code{data.table} which must contain the following columns:
\describe{
  \item{\code{Origin}}{IDs of the origin nodes for each edge.}
  \item{\code{Destination}}{IDs of the destination nodes for each edge.}
  \item{\code{OriginType}}{Types of the origin node for each edge.}
  \item{\code{DestinationType}}{Types of the destination node for each edge.}
  \item{\code{EdgeType}}{Types of each edge.}}}

\item{neighbor_list}{Neighbor reference object constructed by \code{get_neighbor_list()}.}

\item{check}{Should type checking be performed? Default is \code{TRUE}.}

\item{verbose}{Should the intermediate calculations be printed to the console?}
}
\value{
A list with six elements:
\describe{
  \item{Set1}{IDs of node set #1 provided (i.e., \code{set1}).}
  \item{Set2}{ID of node set #2 provided (i.e., \code{set2}).}
  \item{Comparisons}{List of pairwise comparisons.}
  \item{Details}{Intermediate computations produced by \code{get_similarity()}.}
  \item{Similarities}{Pairwise similarity scores.}
  \item{SetSimilarity}{Aggregate set similarity score(s).}}
}
\description{
Given a specified meta-path, similarity metric(s), and aggregation method(s), compute the meta-path based similarity score between two nodes.
}
\seealso{
\code{get_neighbor_list()} for neighbor reference object construction,
\code{get_similarity_function()} for similarity metrics, and
\code{get_aggregation_function()} for aggregation methods. This function is a
wrapper around \code{get_similarity()}.
}
