% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-neighbors.R
\name{get_neighbors_type}
\alias{get_neighbors_type}
\title{Get all node neighbors by type.}
\usage{
get_neighbors_type(root, edge_list_types, edge_list)
}
\arguments{
\item{root}{ID of the root node.}

\item{edge_list_types}{Specified node types to stratify by.}

\item{edge_list}{Edge list as a \code{data.table} which must contain the columns \code{Origin},
\code{Destination}, \code{OriginType}, and \code{DestinationType}.}
}
\value{
Nested list of neighbors grouped by type.
}
\description{
Retrieve all neighbors of a given node, stratified by specified node types.
}
