% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-similarity.R
\name{get_similarity}
\alias{get_similarity}
\title{Compute meta-path based similarity score between two nodes.}
\usage{
get_similarity(
  x,
  y,
  mp,
  metric = c("pc", "npc", "dwpc", "pathsim"),
  node_list,
  edge_list = NULL,
  neighbor_list = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{ID of the origin node.}

\item{y}{ID of the destination node.}

\item{mp}{Meta-path as a vector of node types (e.g., \code{c("TypeA", "TypeB", "TypeC")}).}

\item{metric}{A vector of permissible similarity metrics (e.g., \code{c("npc", "dwpc")}), see defined metrics below:
\describe{
  \item{Path Count}{Specify \code{"pc"}, corresponding to \code{get_pc()}.}
  \item{Normalized Path Count}{Specify \code{"npc"}, corresponding to \code{get_npc()}.}
  \item{Degree-Weighted Path Count}{Specify \code{"dwpc"}, corresponding to \code{get_dwpc()}.}
  \item{PathSim}{Specify \code{"pathsim"}, corresponding to \code{get_pathsim()}.}}}

\item{node_list}{Node list as a \code{data.table} which must contain the following columns:
\describe{
  \item{\code{Node}}{Node IDs (corresponding to either \code{Origin} or \code{Destination} in the edge list).}
  \item{\code{NodeType}}{Node types (corresponding to either \code{OriginType} or
  \code{DestinationType} in the edge list).}}}

\item{edge_list}{Edge list as a \code{data.table} which must contain the following columns:
\describe{
  \item{\code{Origin}}{IDs of the origin nodes for each edge.}
  \item{\code{Destination}}{IDs of the destination nodes for each edge.}
  \item{\code{OriginType}}{Types of the origin node for each edge.}
  \item{\code{DestinationType}}{Types of the destination node for each edge.}}}

\item{neighbor_list}{Neighbor reference object constructed by \code{get_neighbor_list()}.}

\item{verbose}{Should the intermediate calculations be printed to the console?}
}
\value{
A list with six elements:
\describe{
  \item{Origin}{ID of the origin node provided (i.e., \code{x}).}
  \item{Destination}{ID of the destination node provided (i.e., \code{y}).}
  \item{MP}{Meta-path provided (i.e., \code{mp}).}
  \item{OriginPaths}{Paths following the provided meta-path from the origin node (i.e., \code{x}) to all nodes of the
  same type as the destination node (i.e., \code{y}) as a \code{data.table}.}
  \item{DestinationPaths}{Paths following the REVERSE of the provided meta-path from the destination node (i.e., \code{y}) to all nodes of the
  same type as the origin node (i.e., \code{x}) as a \code{data.table}.}
  \item{Similarity}{Computed meta-path based similarity scores by metric as a \code{data.table}.}}
}
\description{
Given a specified meta-path and similarity metric(s), compute the meta-path based similarity score between two nodes.
}
\seealso{
\code{get_neighbor_list()} for neighbor reference object construction and \code{get_similarity_function()} for similarity metrics.
}
