% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-similarity.R
\name{get_similarity}
\alias{get_similarity}
\title{Get node neighbors of specified type using edge list.}
\usage{
get_similarity(
  x,
  y,
  mp,
  metric = c("pc", "npc", "dwpc", "pathsim"),
  edge_list = NULL,
  neighbor_list = NULL,
  verbose = TRUE
)
}
\arguments{
\item{edge_list}{Edge list as a \code{data.table} which must contain the columns \code{Origin},
\code{Destination}, \code{OriginType}, and \code{DestinationType}.}

\item{root}{ID of the root node.}

\item{type}{Desired type of the neighbors.}
}
\value{
Vector of neighbors of a given type.
}
\description{
Retrieve the neighbors of a given node which are of a given type.
}
