% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-types.R
\name{node_types}
\alias{node_types}
\title{Get node types.}
\usage{
node_types(edge_list, verbose = F)
}
\arguments{
\item{edge_list}{Edge list as a \code{data.table} which must contain the following columns:
\describe{
  \item{\code{Origin}}{IDs of the origin nodes for each edge.}
  \item{\code{Destination}}{IDs of the destination nodes for each edge.}
  \item{\code{OriginType}}{Types of the origin node for each edge.}
  \item{\code{DestinationType}}{Types of the destination node for each edge.}
  \item{\code{EdgeType}}{Types of each edge.}}}

\item{verbose}{Should node types be printed to the console?}
}
\value{
List of node types.
}
\description{
Retrieve the node types from an edge list.
}
