% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-degrees.R
\name{search_degrees}
\alias{search_degrees}
\title{Get degree count by type.}
\usage{
search_degrees(
  root,
  type = NA,
  reference_list,
  list_type = c("edge", "neighbor")
)
}
\arguments{
\item{root}{ID of the root node.}

\item{type}{Desired type of the neighbors.}

\item{reference_list}{Either an edge list as a \code{data.table} which must contain the columns \code{Origin},
\code{Destination}, \code{OriginType}, and \code{DestinationType}, or a neighbor reference object constructed
by \code{get_neighbor_list()}.}

\item{list_type}{If an edge list is provided, specify \code{"edge"}. If a neighbor list is provided, specify \code{"neighbor"}.}
}
\value{
Degree count stratified by type.
}
\description{
Compute degree count stratified by node type using either an edge list or neighbor reference object computed by \code{get_neighbor_list()}.
}
\seealso{
\code{get_neighbor_list()} for neighbor reference object construction.
}
